Added sekolah phot

+++ app/views/sekolah/detail.php
@@ -19,7 +19,7 @@
 
   <div class="row my-lg-3">
     <div class="col-lg-7">
-      <h1>Peta</h1>
+      <div id="googleMap"></div>
     </div>
 
     <!-- Info Sekolah -->
@@ -102,7 +102,7 @@
               <td><?= $kelas['kondisi_kelas']; ?></td>
               <td><?= $kelas['jumlah_meja']; ?></td>
               <td><?= $kelas['jumlah_kursi']; ?></td>
-              <td><a class="btn btn-danger btn-sm" href="<?= BASEURL; ?>/sekolah/hapus/<?= $data['sekolah_umum']['id_sekolah']; ?>/<?= $kelas['id_kelas']; ?>" role="button">Hapus</a></td>
+              <td><a class="btn btn-danger btn-sm" href="<?= BASEURL; ?>/sekolah/hapus/<?= $data['sekolah_umum']['id_sekolah']; ?>/<?= $kelas['id_kelas']; ?>" role="button" onclick="return confirm('Hapus Data?');">Hapus</a></td>
             </tr>
           <?php 
             $nomor++;
@@ -121,7 +121,12 @@
         <!-- no, id_kelas, kondisi, jumlah_meja, jumlah_kursi-->
         <thead class="thead-light">
           <tr>
-            <th scope="col" colspan="4">Guru</th>
+            <th scope="col" colspan="3">Guru</th>
+            <!-- <th scope="col">
+              <button type="button" class="btn btn-success btn-sm" data-toggle="modal" data-target="#formModalGuru">
+                Edit Guru
+              </button> 
+            </th> -->
           </tr>
         </thead>
         <tbody>
@@ -129,7 +134,6 @@
             <th scope="col">No. </th>
             <th scope="col">Golongan</th>
             <th scope="col">Banyak</th>
-            <th scope="col">Edit</th>
           </tr>
           <?php 
             $nomor = 1;
@@ -139,11 +143,6 @@
               <td><?= $nomor; ?></td>
               <td><?= $guru['golongan_guru']; ?></td>
               <td><?= $guru['jumlah_guru_gol']; ?></td>
-              <td>
-                <button type="button" class="btn btn-danger btn-sm" data-toggle="modal" data-target="#formModalGuru">
-                  Edit Guru
-                </button> 
-              </td>
             </tr>
           <?php 
             $nomor++;
@@ -162,7 +161,12 @@
         <!-- no, id_kelas, kondisi, jumlah_meja, jumlah_kursi-->
         <thead class="thead-light">
           <tr>
-            <th scope="col" colspan="4">Laboratorium</th>
+            <th scope="col" colspan="3">Laboratorium</th>
+            <!-- <th scope="col">
+              <button type="button" class="btn btn-success btn-sm" data-toggle="modal" data-target="#formModalLabor">
+                Edit Laboratorium
+              </button> 
+            </th> -->
           </tr>
         </thead>
         <tbody>
@@ -170,7 +174,6 @@
             <th scope="col">No. </th>
             <th scope="col">Nama Laboratorium</th>
             <th scope="col">Jumlah</th>
-            <td scope="col">Edit</td>
           </tr>
           <?php 
             $nomor = 1;
@@ -180,11 +183,6 @@
               <td><?= $nomor; ?></td>
               <td><?= $labor['nama_labor']; ?></td>
               <td><?= $labor['jumlah_labor']; ?></td>
-              <td>
-                <button type="button" class="btn btn-danger btn-sm" data-toggle="modal" data-target="#formModalLabor">
-                  Edit Labor
-                </button> 
-              </td>
             </tr>
           <?php 
             $nomor++;
@@ -203,7 +201,12 @@
         <!-- no, id_kelas, kondisi, jumlah_meja, jumlah_kursi-->
         <thead class="thead-light">
           <tr>
-            <th scope="col" colspan="6">Alat</th>
+            <th scope="col" colspan="5">Alat</th>
+            <!-- <th scope="col">
+              <button type="button" class="btn btn-success btn-sm" data-toggle="modal" data-target="#formModalAlat">
+                Edit Alat
+              </button> 
+            </th> -->
           </tr>
         </thead>
         <tbody>
@@ -213,7 +216,6 @@
             <th scope="col">Nama Alat</th>
             <th scope="col">Jumlah Alat Baik</th>
             <th scope="col">Jumlah Alat Kurang Baik</th>
-            <th scope="col">Edit</th>
           </tr>
           <?php 
             $nomor = 1;
@@ -225,11 +227,6 @@
               <td><?= $alat['nama_alat']; ?></td>
               <td><?= $alat['kondisi_alat_baik']; ?></td>
               <td><?= $alat['kondisi_alat_buruk']; ?></td>
-              <td>
-                <button type="button" class="btn btn-danger btn-sm" data-toggle="modal" data-target="#formModalAlat">
-                  Edit Alat
-                </button> 
-              </td>
             </tr>
           <?php 
             $nomor++;
